// Soe.Angular.Spa/src/app/features/manage/sys-edi-message-head/models/sys-edi-message-head.model.ts
export interface ISysEdiMessageRowDTO {
  sysEdiMessageRowId: number;
  sysEdiMessageHeadId: number;
  rowSellerArticleNumber: string;
  rowSellerArticleDescription1: string;
  rowSellerArticleDescription2: string;
  rowSellerRowNumber: string;
  rowBuyerArticleNumber: string;
  rowBuyerRowNumber: string;
  rowDeliveryDate?: Date;
  rowBuyerReference: string;
  rowBuyerObjectId: string;
  rowQuantity?: number;
  rowUnitCode: string;
  rowUnitPrice: string;
  rowDiscountPercent?: number;
  rowDiscountAmount?: number;
  rowDiscountPercent1?: number;
  rowDiscountAmount1?: number;
  rowDiscountPercent2?: number;
  rowDiscountAmount2?: number;
  rowNetAmount?: number;
  rowVatAmount?: number;
  rowVatPercentage?: number;
  externalProductId: number;
  stockCode: string;
}

export class SysEdiMessageRowDTO implements ISysEdiMessageRowDTO {
  sysEdiMessageRowId!: number;
  sysEdiMessageHeadId!: number;
  rowSellerArticleNumber!: string;
  rowSellerArticleDescription1!: string;
  rowSellerArticleDescription2!: string;
  rowSellerRowNumber!: string;
  rowBuyerArticleNumber!: string;
  rowBuyerRowNumber!: string;
  rowDeliveryDate?: Date;
  rowBuyerReference!: string;
  rowBuyerObjectId!: string;
  rowQuantity?: number;
  rowUnitCode!: string;
  rowUnitPrice!: string;
  rowDiscountPercent?: number;
  rowDiscountAmount?: number;
  rowDiscountPercent1?: number;
  rowDiscountAmount1?: number;
  rowDiscountPercent2?: number;
  rowDiscountAmount2?: number;
  rowNetAmount?: number;
  rowVatAmount?: number;
  rowVatPercentage?: number;
  externalProductId!: number;
  stockCode!: string;
}